refactor player module to work with MongoDB
