+++ src/player/schemas/player.schema.ts
@@ -1,0 +1,6 @@
+import * as mongoose from 'mongoose';
+
+export const PlayerSchema = new mongoose.Schema({
+  firstName: String,
+  lastName: String,
+});
